Integrate download ETA calculations in model manager
This commit introduces a new feature that calculates and tracks the Estimated Time of Arrival (ETA) for downloading models in the model manager component. The 'eta' property in the 'DownloadProgress' structure in 'custom.capnp' is changed from 'Float32' to 'UInt32'.

In the 'model_manager.py' file, a new method '_calculate_eta' has been added to perform ETA calculations. An additional dictionary '_download_start_times' has been created to keep track of the start time of each model download. The ETA is calculated every time a portion of the model file is downloaded, and it gets updated in the 'DownloadProgress' structure. Finally, the start time is cleared after the download completes.

In 'model_manager_audit.py', an additional check is added to only print downloadProgress for the downloads currently in progress.

+++ cereal/custom.capnp
@@ -44,7 +44,7 @@ struct ModelManagerSP @0xaedffd8f31e7b55d {
     struct DownloadProgress {
         status @0 :DownloadStatus;
         progress @1 :Float32;
-        eta @2 :Float32;
+        eta @2 :UInt32;
     }
 
     struct ModelBundle {
